[Download] First successful download and import implementation

+++ src/main/kotlin/com/github/schaka/naviseerr/download_client/slskd/dto/SearchFile.kt
@@ -2,9 +2,9 @@ package com.github.schaka.naviseerr.download_client.slskd.dto
 
 data class SearchFile(
     val filename: String,
-    val size: Int,
+    val size: Long,
     val extension: String,
-    val length: Long? = null,
+    val length: Long? = null, // seconds the song lasts
     val bitRate: Int? = null,
     val bitDepth: Int? = null,
     val sampleRate: Int? = null,
